Update watch.js to watch and respond to deleted folders.
Noticed that deleting folders weren't reflected in the design manager, which left some empty folders lying around when reorganizing within the local environment. This seemed like the best fix, tested it locally for this use case but not sure if there are other use cases that this might conflict with.